requirements.txt
fastapi
uvicorn[standard]
httpx
aiosqlite
matplotlib
pandas
python-jose
python-multipart
python-dotenv
pydantic-settings